Remove 'use all CPUs' option, not feasible
As the Docker container that executes `run_gpl.py` already has the CPU
number configured, it is not possible to use 'extra' CPUs that SPOT may
have (over-)provisioned. There could be some way to allow AWS to update
the Docker container CPU count but I won't bother with investigating
this for now.